Update cart UI

+++ src/java/model/order/Cart.java
@@ -10,33 +10,32 @@ package model.order;
  * @author Admin
  */
 public class Cart {
-
     private int id;
     private float totalPrice;
 
-    public Cart() {
-
-    }
-
-    public Cart(int id, int Quantity, float totalPrice) {
-        this.id = id;
+   public Cart(){
+       
+   }
+    public Cart(int id,int Quantity, float totalPrice) {
+        this.id=id;
         this.totalPrice = totalPrice;
     }
-
     public int getId() {
         return id;
     }
-
     public void setId(int id) {
         this.id = id;
     }
+
 
     public void setTotalPrice(float totalPrice) {
         this.totalPrice = totalPrice;
     }
 
-    public double getTotalPrice() {
+
+    public float getTotalPrice() {
         return totalPrice;
     }
-
+    
+    
 }
